backend/requirements.txt
# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0

# HTTP & Async
httpx>=0.27.0
requests>=2.32.0
aiohttp>=3.9.0

# ===== FIXED CORE: Let pip resolve these together =====
protobuf>=5.26.1
googleapis-common-protos>=1.65.0
grpcio>=1.68.1
grpcio-status>=1.68.1
google-api-core>=2.23.0

# LLM and AI integrations (Updated for 2026 compatibility)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=4.2.0
google-generativeai>=0.8.3
# ADDED BY JANA FOR GROQ SUPPORT IN THE COORDINATOR AGENT
langchain_groq
# google-ai-generativelanguage is now handled by the consolidated genai SDK

# Google API dependencies
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Memory & Database integrations
langgraph-checkpoint-sqlite>=2.0.0
langgraph-checkpoint-mongodb>=0.3.0
langchain-chroma>=0.1.0
langchain-mongodb>=0.2.0

# MongoDB drivers
pymongo>=4.10.0
motor>=3.6.0
dnspython>=2.7.0

# Mem0 and Vector DB
mem0ai>=1.0.2
qdrant-client>=1.12.0
openai>=1.50.0

# LangGraph core
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0

# Data handling (Pinned to avoid binary conflicts)
pydantic>=2.10.0
pydantic-core>=2.27.0
typing-extensions>=4.12.0
numpy>=1.26.0

# Utilities
tenacity>=8.5.0
langsmith>=0.1.0
jsonschema>=4.20.0
pyyaml>=6.0.0

# Windows-specific
pywin32>=226; sys_platform == 'win32'

sentence-transformers

# OmniParser dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
opencv-python>=4.7.0
ultralytics>=8.0.0  # For YOLO model